comlpiler warn

+++ src/Etterna/Globals/MinaCalc.cpp
@@ -306,7 +306,7 @@ Calc::CalcMain(const vector<NoteInfo>& NoteInfo,
 	float shortstamdownscaler = CalcClamp(
 	  0.9f + (0.1f * (NoteInfo.back().rowTime - 150.f) / 150.f), 0.9f, 1.f);
 
-	static const float fo_rizzy = 1;
+	static const int fo_rizzy = 1;
 	vector<vector<float>> the_hizzle_dizzles(fo_rizzy);
 	for (int WHAT_IS_EVEN_HAPPEN_THE_BOMB = 0;
 		 WHAT_IS_EVEN_HAPPEN_THE_BOMB < fo_rizzy;
